requirements.txt
asttokens           ~=2.4.1
certifi             ~=2024.8.30
charset-normalizer  ~=3.4.0
colorama            ~=0.4.6
contourpy           ~=1.3.0
coverage            ~=7.6.1
cycler              ~=0.12.1
decorator           ~=5.1.1
exceptiongroup      ~=1.2.0
executing           ~=2.1.0
fonttools           ~=4.54.1
idna                ~=3.10
importlib_resources ~=6.4.5
iniconfig           ~=1.1.1
ipython             ~=8.18.1
jedi                ~=0.19.1
joblib              ~=1.4.2
jupyter             ~=1.1.1
kiwisolver          ~=1.4.7
matplotlib          ~=3.9.2
matplotlib-inline   ~=0.1.7
mplfinance          ~=0.12.10b0
numpy               ~=2.0.2
packaging           ~=24.1
pandas              ~=2.2.3
parso               ~=0.8.4
pillow              ~=11.0.0
pip                 ~=24.2
pluggy              ~=1.0.0
prompt_toolkit      ~=3.0.48
psutil              ~=6.1.0
pure_eval           ~=0.2.3
Pygments            ~=2.18.0
PyJWT               ~=2.9.0
pyparsing           ~=3.2.0
pytest              ~=7.4.4
python-dateutil     ~=2.9.0.post0
python-dotenv       ~=1.0.1
pytz                ~=2024.2
pyupbit             ~=0.2.34
requests            ~=2.32.3
scikit-learn        ~=1.5.2
scipy               ~=1.13.1
setuptools          ~=75.1.0
six                 ~=1.16.0
smtm                ~=0.1
stack-data          ~=0.6.3
threadpoolctl       ~=3.5.0
tomli               ~=2.0.1
traitlets           ~=5.14.3
typing_extensions   ~=4.12.2
tzdata              ~=2024.2
urllib3             ~=2.2.3
wcwidth             ~=0.2.13
websockets          ~=13.1
wheel               ~=0.44.0
zipp                ~=3.20.2
